+++ src/users.rs
@@ -1,134 +1,14 @@
 //user.rs
 use crate::{
-    response::{Response, MalCommonResponse},
+    response::Response,
     enums::users::{Gender, UserHistoryType},
     structs::{
-        anime::Anime, 
-        character::Character, 
-        manga::Manga, 
-        people::Person, 
         recommendation::Recommendation, 
-        reviews::Review
-    }, utils::{ExternalEntry, Images}, JikanClient, JikanError
+        reviews::Review,
+        users::{User, UserStats, UserUpdates, Friend, UserHistory, UserFavorite, UserAbout},
+        clubs::Club,
+    }, utils::ExternalEntry, JikanClient, JikanError
 };
-use serde::{Deserialize, Serialize};
-
-#[derive(Debug, Clone, Serialize, Deserialize)]
-pub struct UserById {
-    pub url: String,
-    pub username: String,
-}
-
-#[derive(Debug, Clone, Serialize, Deserialize)]
-pub struct UserAnimeInfo {
-    pub days_watched: f32,
-    pub mean_score: f32,
-    pub watching: i32,
-    pub completed: i32,
-    pub on_hold: i32,
-    pub dropped: i32,
-    pub plan_to_watch: i32,
-    pub total_entries: i32,
-    pub rewatched: i32,
-    pub episodes_watched: i32,
-}
-
-#[derive(Debug, Clone, Serialize, Deserialize)]
-pub struct UserMangaInfo {
-    pub days_read: f32,
-    pub mean_score: f32,
-    pub reading: i32,
-    pub completed: i32,
-    pub on_hold: i32,
-    pub dropped: i32,
-    pub plan_to_read: i32,
-    pub total_entries: i32,
-    pub reread: i32,
-    pub chapters_read: i32,
-    pub volumes_read: i32,
-}
-
-#[derive(Debug, Clone, Serialize, Deserialize)]
-pub struct UserStats {
-    pub anime: UserAnimeInfo,
-    pub manga: UserMangaInfo,
-}
-
-#[derive(Debug, Clone, Serialize, Deserialize)]
-pub struct UserAbout {
-    pub about: Option<String>,
-}
-
-#[derive(Debug, Clone, Serialize, Deserialize)]
-pub struct UserClub {
-    pub mal_id: u32,
-    pub name: String,
-    pub url: String,
-}
-
-#[derive(Debug, Clone, Serialize, Deserialize)]
-pub struct UserAnimeUpdates {
-    pub entry: Anime,
-    pub score: Option<i32>,
-    pub status: String,
-    pub episodes_seen: Option<u32>,
-    pub episodes_total: Option<u32>,
-    pub date: String,
-}
-
-#[derive(Debug, Clone, Serialize, Deserialize)]
-pub struct UserMangaUpdates {
-    pub entry: Manga,
-    pub score: Option<i32>,
-    pub status: String,
-    pub chapters_read: Option<u32>,
-    pub chapters_total: Option<u32>,
-    pub volumes_read: Option<u32>,
-    pub volumes_total: Option<u32>,
-    pub date: String,
-}
-
-#[derive(Debug, Clone, Serialize, Deserialize)]
-pub struct Friend {
-    user: User,
-    last_online: Option<String>,
-    friends_since: Option<String>,
-}
-
-#[derive(Debug, Clone, Serialize, Deserialize)]
-pub struct UserHistory {
-    pub entry: MalCommonResponse,
-    pub increment: Option<i32>,
-    pub date: Option<String>,
-}
-
-#[derive(Debug, Clone, Serialize, Deserialize)]
-pub struct UserFavorite {
-    pub anime: Vec<Anime>,
-    pub manga: Vec<Manga>,
-    pub characters: Vec<Character>,
-    pub people: Vec<Person>,
-}
-
-#[derive(Debug, Clone, Serialize, Deserialize)]
-pub struct UserUpdates {
-    anime: Vec<UserAnimeUpdates>,
-    manga: Vec<UserMangaUpdates>,
-}
-
-#[derive(Debug, Clone, Serialize, Deserialize)]
-pub struct User {
-    pub username: String,
-    pub url: String,
-    pub images: Option<Images>,
-}
-
-#[derive(Debug, Clone, Serialize, Deserialize)]
-pub struct ReviewUser {
-    pub url: String,
-    pub username: String,
-    pub images: Images,
-}
 
 pub struct GetUsersParams {
     pub page: Option<String>,
@@ -192,7 +72,7 @@ impl JikanClient {
         self.get(&format!("/users/{}", query)).await
     }
 
-    pub async fn get_user_by_id(&self, id: i32) -> Result<Response<UserById>, JikanError> {
+    pub async fn get_user_by_id(&self, id: i32) -> Result<Response<User>, JikanError> {
         //Maybe handle this better?
         self.get(&format!("/users/userbyid/{}", id)).await
     }
@@ -252,7 +132,7 @@ impl JikanClient {
         &self,
         username: &str,
         page: Option<u32>,
-    ) -> Result<Response<Vec<UserClub>>, JikanError> {
+    ) -> Result<Response<Vec<Club>>, JikanError> {
         let mut params = String::new();
         if let Some(p) = page {
             params = format!("?page={}", p);
